% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_link}
\alias{build_link}
\title{Builds the link to be passed to the API}
\usage{
build_link(path_elements, query_elements = NULL, before = NULL, after = NULL)
}
\arguments{
\item{path_elements}{Specifies elements of the path}

\item{query_elements}{Specifies elements of the query}

\item{before}{The fullname of an item serving as anchor in the request.
Items before this item in the listing are returned.}

\item{after}{The fullname of an item serving as anchor in the request.
Items after this item in the listing are returned.}
}
\value{
A string with the link to be passed to the API.
}
\description{
Builds the link to be passed to the API
}
\keyword{internal}
