% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_args}
\alias{check_args}
\title{Checks whether valid arguments are supplied to a function}
\usage{
check_args(default_arg = NULL)
}
\arguments{
\item{default_arg}{Specifies an argument that has to be set by default.}
}
\value{
The function returns NULL if the arguments are valid.
}
\description{
This is an internal function.
}
\keyword{internal}
