% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_credentials}
\alias{check_credentials}
\title{Checks credentials}
\usage{
check_credentials(cred)
}
\arguments{
\item{cred}{The credential to be checked}
}
\value{
The function returns NULL if the credential is valid.
}
\description{
Checks whether credentials such as username, password, client id, 
client secret are valid strings
}
\keyword{internal}
