% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_token}
\alias{check_token}
\title{Checks token for their validity}
\usage{
check_token(token, scope)
}
\arguments{
\item{token}{The token to check}

\item{scope}{The scope of the token}
}
\value{
The function returns NULL if the token is valid.
}
\description{
This is an internal function.
}
\keyword{internal}
