% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-functions.R
\name{clean_posts}
\alias{clean_posts}
\title{Clean and shape subreddit data}
\usage{
clean_posts(x)
}
\value{
A tibble of subreddit posts and associated metadata
}
\description{
Write more about \code{clean_posts()} here.
}
\examples{

\dontrun{

example_posts <- get_posts("education")
example_posts <- clean_posts(example_posts)
min(example_posts$post_date_time)  # Returns date-time of most recent post
}
}
\seealso{
Read more about \href{https://github.com/reddit-archive/reddit/wiki/API}{OAuth2 authentication for the Reddit API}
  and \href{https://www.reddit.com/prefs/apps}{creating your own appication key}.
}
