% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_comments}
\alias{get_comments}
\title{Get comments from a specified subreddit}
\usage{
get_comments(subreddit, accesstoken, limit = 100, before = NULL,
  after = NULL, verbose = FALSE)
}
\arguments{
\item{subreddit}{The name of the subreddit from which comments are requested}

\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"read"}.}

\item{limit}{The maximum number of comments to return. Must be a number
between 1 and 100.}

\item{before}{The fullname of a comment serving as anchor in the request.
Comments before this comment in the listing are returned.}

\item{after}{The fullname of a comment serving as anchor in the request.
Comments after this comment in the listing are returned.}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}
}
\value{
A dataframe with new comments from a specified subreddit.
}
\description{
Get comments from a specified subreddit
}
