% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_reddit}
\alias{get_reddit}
\title{GET method for the reddit API}
\usage{
get_reddit(accesstoken, link, verbose, retry)
}
\arguments{
\item{accesstoken}{The token for the GET method}

\item{link}{The link used in the GET method}

\item{verbose}{Boolean. A logical flag whether information about the data 
extraction should be printed to the console.}

\item{retry}{Boolean. A logical flag whether a failed api request should be 
retried if it fails.Requests will be tried up to three times with varying 
time intervals between requests.}
}
\value{
A JSON element
}
\description{
This is an internal function.
}
\keyword{internal}
