% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_submissions}
\alias{get_submissions}
\title{Get submissions for a specified user}
\usage{
get_submissions(
  user,
  accesstoken,
  sort = "new",
  time = NULL,
  limit = 100,
  before = NULL,
  after = NULL,
  output = c("df", "json", "all"),
  verbose = FALSE,
  retry = FALSE
)
}
\arguments{
\item{user}{The username of the user for which submissions are requested.}

\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"history"}.}

\item{sort}{The order in which the submissions of the user should be ordered
when accessing them. Possible values are:
\itemize{
  \item \code{new}: Sorts posts by the time when they were created
  \item \code{hot}: Sorts posts by posts currently trending
  \item \code{controversial}: Sorts posts by number of upvotes and downvotes
  \item \code{top}: Most upvoted posts in a certain timeframe. Timeframe can
  be specified with \code{time}.
}}

\item{time}{The timeframe for the \code{sort} order \code{controversial} and
\code{top}. Possible values are:
\itemize{
  \item \code{hour}
  \item \code{day}
  \item \code{week}
  \item \code{month}
  \item \code{year}
  \item \code{all time}
}}

\item{limit}{The maximum number of submissions to return. Must be a number
between 1 and 100.}

\item{before}{The fullname of a post serving as anchor in the request.
Posts before this post in the listing are returned.}

\item{after}{The fullname of a post serving as anchor in the request.
Posts after this post in the listing are returned.}

\item{output}{What the function should return. Choose \code{json} for an 
unparsed json object, \code{df} for a parsed object in form of a dataframe, 
and \code{all} for a list containg the json object, a dataframe, and the 
before and after anchors (if those exist for the endpoint).}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}

\item{retry}{A logical flag whether a failed api request should be retried.
Requests will be tried up to three times with varying time intervals between
requests.}
}
\value{
A dataframe of posts for a specified user.
}
\description{
Submissions are posts made by one user. Submissions can be requested for a
specific timeframe or sort order.
}
\details{
More info at \url{https://www.reddit.com/dev/api/}
}
\examples{
\dontrun{
history_token <- get_token(scope = "history",
                           useragent = useragent,
                           username = username,
                           password = password,
                           client_id = client_id,
                           client_secret = client_secret)

submissions <- get_submissions(user = "_KeanuReeves",
                              accesstoken = history_token,
                              sort = "top", time = "all",
                              limit = 10)
}
}
\seealso{
\code{\link{get_user}}
}
