% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_subreddit_info}
\alias{get_subreddit_info}
\title{Get basic information for a specified subreddit.}
\usage{
get_subreddit_info(
  subreddit,
  type = c("info", "moderators", "rules"),
  accesstoken,
  output = c("df", "json", "all"),
  verbose = FALSE,
  retry = FALSE
)
}
\arguments{
\item{subreddit}{Name of the subreddit for which info is requested.}

\item{type}{Which info is requested. Possible values are \code{info} for
information about the subreddit from the sidebar, \code{moderators} for
the moderators of the subreddit and \code{rules} for the rules of the
subreddit.}

\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"read"}.}

\item{output}{What the function should return. Choose \code{json} for an 
unparsed json object, \code{df} for a parsed object in form of a dataframe, 
and \code{all} for a list containg the json object, a dataframe, and the 
before and after anchors (if those exist for the endpoint).}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}

\item{retry}{A logical flag whether a failed api request should be retried.
Requests will be tried up to three times with varying time intervals between
requests.}
}
\value{
Basic information about a subreddit from the
sidebar or the moderators or the rules of the subreddit as a dataframe, json 
object or a list containing both.
}
\description{
Get basic information for a specified subreddit.
}
\details{
More info at \url{https://www.reddit.com/dev/api/}
}
\examples{
\dontrun{
read_token <- get_token(scope = "read",
                        useragent = my_useragent,
                        username = my_username,
                        password = my_password,
                        client_id = my_client_id,
                        client_secret = my_client_secret)

#Get sidebar for subreddit "soccer"
sub_info <- get_subreddit_info(subreddit = "soccer",
                               type = "info",
                               accesstoken = read_token)

#Get moderators for subreddit "soccer"
moderators <- get_subreddit_info(subreddit = "soccer",
                                type = "moderators",
                                accesstoken = read_token)
}
}
