% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_subreddits}
\alias{get_subreddits}
\title{Get a dataframe of subreddits}
\usage{
get_subreddits(
  type = c("popular", "new", "default", "search"),
  accesstoken,
  query = NULL,
  sort = c("relevance", "activity"),
  limit = 100,
  after = NULL,
  before = NULL,
  output = c("df", "json", "all"),
  verbose = FALSE,
  retry = FALSE
)
}
\arguments{
\item{type}{The type of list that is requested. Possible values are:
\itemize{
  \item \code{popular} Subreddits that are popular right now
  \item \code{new} Newly created subreddits
  \item \code{default} Default subreddits that users are subscribed to when
  they sign up for reddit.
   \item \code{search} Search for subreddits with a query.
  }}

\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"read"}.}

\item{query}{Search terms for type \code{search}. Must be less than 512
characters.}

\item{sort}{The order in which results from type \code{search} should be
sorted. Possible values are \code{"relevance"} and \code{"activity"}.}

\item{limit}{The maximum number of subreddits to return. Must be a number
between 1 and 100.}

\item{after}{The fullname of an item serving as anchor in the request.
Items after this item in the listing are returned.}

\item{before}{The fullname of an item serving as anchor in the
request. Items before this item in the listing are returned.}

\item{output}{What the function should return. Choose \code{json} for an 
unparsed json object, \code{df} for a parsed object in form of a dataframe, 
and \code{all} for a list containg the json object, a dataframe, and the 
before and after anchors (if those exist for the endpoint).}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}

\item{retry}{A logical flag whether a failed api request should be retried.
Requests will be tried up to three times with varying time intervals between
requests.}
}
\value{
New or popular subreddits as a dataframe, json object or a list 
containing both.
}
\description{
Subreddits can be sorted by creation date or popularity. Also
allows searching for subreddits with search queries.
}
\details{
More info at \url{https://www.reddit.com/dev/api/}
}
\examples{
\dontrun{
read_token <- get_token(scope = "read",
                        useragent = my_useragent,
                        username = my_username,
                        password = my_password,
                        client_id = my_client_id,
                        client_secret = my_client_secret)
                        
#type = "popular" will show subreddits that are popular right now
popular_subs <- get_subreddits(type = "popular",
                              limit = 10,
                              accesstoken = read_token)

#Searches for subreddits can be performed with type "search"
search_subs <- get_subreddits(type = "search",
                             accesstoken = read_token,
                             limit = 10,
                             query = "soccer",
                             sort = "activity")
}
}
