% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_thing}
\alias{get_thing}
\title{Get a comment or post based on its unique id}
\usage{
get_thing(
  subreddit,
  id,
  type = c("comment", "post"),
  accesstoken = NULL,
  output = c("df", "json", "all"),
  verbose = FALSE,
  retry = FALSE
)
}
\arguments{
\item{subreddit}{The subreddit of the thing}

\item{id}{The unique id of the requested thing. Reddit IDs are in base 36.}

\item{type}{What type of thing the id is. Must be \code{comment} or 
\code{post}.}

\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"read"}.}

\item{output}{What the function should return. Choose \code{json} for an 
unparsed json object, \code{df} for a parsed object in form of a dataframe, 
and \code{all} for a list containg the json object, a dataframe, and the 
before and after anchors (if those exist for the endpoint).}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}

\item{retry}{A logical flag whether a failed api request should be retried.
Requests will be tried up to three times with varying time intervals between
requests.}
}
\value{
The requested thing as a dataframe, json object or a list containing 
both.
}
\description{
Get a comment or post based on its unique id
}
\details{
More info at \url{https://www.reddit.com/dev/api/}
}
\examples{
\dontrun{
read_token <- get_token(scope = "read",
                        useragent = my_useragent,
                        username = my_username,
                        password = my_password,
                        client_id = my_client_id,
                        client_secret = my_client_secret)
                        

comm <- get_comments(subreddit = "askreddit",
                      accesstoken = read_token,
                      output = "df",
                      limit = 1)

#Look up the post this comment was made to
parent_post <- get_thing(subreddit = "askreddit",
                         id = comm$parent_id,
                         item_type = "post",
                         accesstoken = read_token)
}
}
