% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_user_comments}
\alias{get_user_comments}
\title{Get comments from a specified user}
\usage{
get_user_comments(
  user,
  accesstoken,
  sort = "new",
  time = NULL,
  limit = 100,
  after = NULL,
  before = NULL,
  verbose = FALSE,
  retry = TRUE
)
}
\arguments{
\item{user}{The username of the user for which comments are requested.}

\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"history"}.}

\item{sort}{The order in which the submissions of the user should be ordered
when accessing them. Possible values are:
\itemize{
  \item \code{new}: Returns newest comments
  \item \code{hot}: Returns highly upvoted comments currently
  \item \code{controversial}: Returns comments that received a lot of upvotes
  and downvotes
  \item \code{top}: Returns most upvoted comments in a certain timeframe.
  Timeframe can be specified with \code{time}.
}}

\item{time}{The timeframe for the \code{sort} order \code{controversial} and
\code{top}. Possible values are:
\itemize{
  \item \code{hour}
  \item \code{day}
  \item \code{week}
  \item \code{month}
  \item \code{year}
  \item \code{all time}
}}

\item{limit}{The maximum number of user comments to return. Must be a number
between 1 and 100.}

\item{after}{The fullname of a comment serving as anchor in the request.
Comments after this comment in the listing are returned.}

\item{before}{The fullname of a comment serving as anchor in the request.
Comments before this comment in the listing are returned.}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}

\item{retry}{A logical flag whether a failed api request should be retried.
Requests will be tried up to three times with varying time intervals between
requests.}
}
\value{
A dataframe of comments for the specified user.
}
\description{
Get comments from a specified user
}
\examples{
\dontrun{
#Get 10 comments made by Barack Obama
obama_comments <- get_user_comments(user = "PresidentObama",
                                    accesstoken = his_token,
                                    sort = "top", time = "all",
                                    limit = 10)
}
}
