% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_users}
\alias{get_users}
\title{Get a list of user accounts}
\usage{
get_users(type = c("popular", "new"), accesstoken, limit = 100,
  after = NULL, before = NULL, verbose = FALSE)
}
\arguments{
\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"read"}.}

\item{limit}{The maximum number of users to return. Must be a number
between 1 and 100.}

\item{after}{The fullname of an item serving as anchor in the request.
Items after this item in the listing are returned.}

\item{before}{The fullname of an item serving as anchor in the
request. Items before this item in the listing are returned.}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}
}
\value{
A dataframe of users.
}
\description{
Get a list of user accounts
}
