% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_request}
\alias{parse_request}
\title{Parse the request from the reddit API}
\usage{
parse_request(request, verbose, output, after_before = TRUE)
}
\arguments{
\item{request}{The request made to the reddit API}

\item{verbose}{Boolean. A logical flag whether information about the data 
extraction should be printed to the console.}

\item{output}{What the function should return. Choose \code{json} for an 
unparsed json object, \code{df} for a parsed object in form of a dataframe, 
and \code{all} for a list containg the json object, a dataframe, and the 
before and after anchors.}

\item{after_before}{Boolean. A logical flag whether \code{output = "all"} 
should return the before and after anchors.}
}
\value{
A dataframe or vector.
}
\description{
The function parses the JSON object of the request of the reddit API and 
turns it into a dataframe or vector if so desired.
}
\keyword{internal}
