% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_comments}
\alias{get_comments}
\title{Get comments from a specified subreddit}
\usage{
get_comments(
  subreddit,
  accesstoken,
  limit = 100,
  before = NULL,
  after = NULL,
  verbose = FALSE,
  retry = FALSE
)
}
\arguments{
\item{subreddit}{The name of the subreddit from which comments are requested}

\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"read"}.}

\item{limit}{The maximum number of comments to return. Must be a number
between 1 and 100.}

\item{before}{The fullname of a comment serving as anchor in the request.
Comments before this comment in the listing are returned.}

\item{after}{The fullname of a comment serving as anchor in the request.
Comments after this comment in the listing are returned.}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}

\item{retry}{A logical flag whether a failed api request should be retried.
Requests will be tried up to three times with varying time intervals between
requests.}
}
\value{
A dataframe with new comments from a specified subreddit.
}
\description{
Get comments from a specified subreddit
}
\examples{
\dontrun{

#10 new comments from subreddit askreddit
comms <- get_comments(subreddit = "askreddit",
                      accesstoken = read_token,
                      limit = 10)
}
}
