% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-functions.R
\name{get_posts}
\alias{get_posts}
\title{Retrieve subreddit posts and bring into R}
\usage{
get_posts(
  subreddit,
  max_posts = 100,
  starting_point = "",
  starting_data = NULL
)
}
\value{
A tibble of subreddit posts and associated metadata
}
\description{
Write more about \code{get_posts()} here.
}
\examples{

\dontrun{

get_posts("education")
get_posts("instructionaldesign", max_posts = 10)
}
}
\seealso{
Read more about \href{https://github.com/reddit-archive/reddit/wiki/API}{OAuth2 authentication for the Reddit API}
  and \href{https://www.reddit.com/prefs/apps}{creating your own appication key}.
}
