% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_posts}
\alias{get_posts}
\title{Get posts from a specified subreddit}
\usage{
get_posts(subreddit, accesstoken, sort = "new", limit = 100,
  time = NULL, link_flair = FALSE, author_flair = FALSE,
  user_reports = FALSE, secure_media = FALSE, awards = FALSE,
  after = NULL, before = NULL, verbose = TRUE)
}
\arguments{
\item{subreddit}{The name of the subreddit from which posts are requested.}

\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"read"}.}

\item{sort}{The order in which the posts in the subreddit should be ordered
when accessing them. Possible values are:
\itemize{
  \item \code{new}: Sorts posts by the time when they were created
  \item \code{hot}: Sorts posts by posts currently
  \item \code{controversial}:
  \item \code{random}:
  \item \code{rising}:
  \item \code{top}: Most upvoted posts in a certain timeframe. Timeframe can
  be specified with \code{time}.
}}

\item{limit}{The maximum number of posts to return. Must be a number between
1 and 100.}

\item{time}{The timeframe for the \code{sort} order \code{controversial} and
\code{top}. Possible values are:
\itemize{
  \item \code{hour}
  \item \code{day}
  \item \code{week}
  \item \code{month}
  \item \code{year}
  \item \code{all time}
}}

\item{after}{The fullname of a post serving as anchor in the request.
Posts after this post in the listing are returned.}

\item{before}{The fullname of a post serving as anchor in the request.
Posts before this post in the listing are returned.}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}
}
\value{
A dataframe containg the requested posts
}
\description{
Get posts from a specified subreddit
}
\details{
More info at \url{https://www.reddit.com/dev/api/}
}
