% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_submissions}
\alias{get_submissions}
\title{Get submissions for a specified user}
\usage{
get_submissions(user, accesstoken, sort = "new", time = NULL,
  limit = 100, before = NULL, after = NULL, verbose = FALSE,
  retry = TRUE)
}
\arguments{
\item{user}{The username of the user for which submissions are requested.}

\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"history"}.}

\item{sort}{The order in which the submissions of the user should be ordered
when accessing them. Possible values are:
\itemize{
  \item \code{new}: Sorts posts by the time when they were created
  \item \code{hot}: Sorts posts by posts currently trending
  \item \code{controversial}: Sorts posts by number of upvotes and downvotes
  \item \code{top}: Most upvoted posts in a certain timeframe. Timeframe can
  be specified with \code{time}.
}}

\item{time}{The timeframe for the \code{sort} order \code{controversial} and
\code{top}. Possible values are:
\itemize{
  \item \code{hour}
  \item \code{day}
  \item \code{week}
  \item \code{month}
  \item \code{year}
  \item \code{all time}
}}

\item{limit}{The maximum number of submissions to return. Must be a number
between 1 and 100.}

\item{before}{The fullname of a post serving as anchor in the request.
Posts before this post in the listing are returned.}

\item{after}{The fullname of a post serving as anchor in the request.
Posts after this post in the listing are returned.}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}

\item{retry}{A logical flag whether a failed api request should be retried.
Requests will be tried up to three times with varying time intervals between
requests.}
}
\value{
A dataframe of posts for a specified user.
}
\description{
Get submissions for a specified user
}
\details{
More info at \url{https://www.reddit.com/dev/api/}
}
