% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_subreddit_info}
\alias{get_subreddit_info}
\title{Get basic information for a specified subreddit.}
\usage{
get_subreddit_info(subreddit, type = c("info", "moderators", "rules"),
  accesstoken, verbose = FALSE, retry = TRUE)
}
\arguments{
\item{subreddit}{Name of the subreddit for which info is requested.}

\item{type}{Which info is requested. Possible values are \code{info} for
information about the subreddit from the sidebar, \code{moderators} for
the moderators of the subreddit and \code{rules} for the rules of the
subreddit.}

\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"read"}.}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}

\item{retry}{A logical flag whether a failed api request should be retried.
Requests will be tried up to three times with varying time intervals between
requests.}
}
\value{
A dataframe with basic information about a subreddit from the
sidebar or a dataframe with the moderators or the rules of the subreddif.
}
\description{
Get basic information for a specified subreddit.
}
\examples{
\dontrun{
read_token <- get_token(scope = "read",
                                      useragent = useragent,
                                      username = username,
                                      password = password)

sub_info <- get_subreddit_info(subreddit = "soccer",
                               accesstoken = read_token,
                               verbose = FALSE)
                               }
}
