% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_subreddits}
\alias{get_subreddits}
\title{Get a list of subreddits}
\usage{
get_subreddits(type = c("popular", "new", "default", "search"),
  accesstoken, query = NULL, sort = NULL, limit = 100,
  after = NULL, before = NULL, verbose = FALSE)
}
\arguments{
\item{type}{The type of list that is requested. Possible values are:
\itemize{
  \item \code{popular} Subreddits that are popular right now
  \item \code{new} Newly created subreddits
  \item \code{default} Default subreddits that users are subscribed to when
  they sign up for reddit.
   \item \code{search} Search for subreddits with a query.
  }}

\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"read"}.}

\item{query}{Search terms for type \code{search}. Must be less than 512
characters.}

\item{sort}{The order in which results from type \code{search} should be
sorted. Possible values are \code{relevance} and \code{activity}.}

\item{limit}{The maximum number of subreddits to return. Must be a number
between 1 and 100.}

\item{after}{The fullname of an item serving as anchor in the request.
Items after this item in the listing are returned.}

\item{before}{The fullname of an item serving as anchor in the
request. Items before this item in the listing are returned.}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}
}
\value{
A dataframe of subreddits.
}
\description{
Get a list of subreddits
}
