% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_token}
\alias{get_token}
\title{Generate an access token for the Reddit API}
\usage{
get_token(
  scope = c("identity", "read", "history", "wikiread"),
  useragent,
  username = NULL,
  password = NULL,
  client_id = NULL,
  client_secret = NULL
)
}
\arguments{
\item{scope}{The scope of what you want to access. Must be one of these:
identity, read, history, wikiread}

\item{useragent}{The useragent for the device accessing the API}

\item{username}{Username for the Reddit profile accessing the API. If left
empty, it will attempt to read from the environment. Can be set with
\code{Sys.setenv(REDDIT_API_USERNAME = "")}}

\item{password}{Password for the Reddit profile accessing the API. If left
empty, it will attempt to read from the environment. Can be set with
\code{Sys.setenv(REDDIT_API_PASSWORD = "")}}

\item{client_id}{The id of the app through which the API is accessed.If left
empty, it will attempt to read from the environment. Can be set with
\code{Sys.setenv(REDDIT_API_CLIENT_ID = "")}}

\item{client_secret}{The client secret of the app through which the API is
accessed. If left empty, it will attempt to read from the environment. Can be
set with\code{Sys.setenv(REDDIT_API_SECRET = "")}}
}
\value{
A list containing the access token, the time until it expires, the
scope, the time it was generated and the useragent
}
\description{
Generate an access token for the Reddit API
}
\details{

}
