% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_token}
\alias{get_token}
\title{Generate an access token for the Reddit API}
\usage{
get_token(scope = c("identity", "read", "history", "wikiread"),
  useragent, username, password)
}
\arguments{
\item{scope}{The scope of what you want to access. Must be one of these:
identity, read, history, wikiread}

\item{useragent}{The useragent for the device accessing the API}

\item{username}{Username for the Reddit profile accessing the API}

\item{password}{Password for the Reddit profile accessing the API}
}
\value{
A list containing the access token, the time until it expires, the
scope, the time it was generated and the useragent
}
\description{
Generate an access token for the Reddit API
}
\details{
More info at \url{https://www.reddit.com/dev/api/}
}
