% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_trophies}
\alias{get_trophies}
\title{Get trophies for a specified user}
\usage{
get_trophies(user, accesstoken, verbose = FALSE, retry = TRUE)
}
\arguments{
\item{user}{The username of the user}

\item{accesstoken}{The access token required to access the endpoint. Scope of
the acces token must be \code{"read"}.}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}

\item{retry}{A logical flag whether a failed api request should be retried.
Requests will be tried up to three times with varying time intervals between
requests.}
}
\value{
A dataframe with all trophies a specified user has received.
}
\description{
Get trophies for a specified user
}
\seealso{
\code{\link{get_user_info}}
}
