% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_user_info}
\alias{get_user_info}
\title{Get basic information about account for a specified user}
\usage{
get_user_info(user, accesstoken, verbose = FALSE)
}
\arguments{
\item{user}{Username of the Reddit user that is requested.}

\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"read"}.}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}
}
\value{
A dataframe with information about a specified user.
}
\description{
Get basic information about account for a specified user
}
