% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_user_info}
\alias{get_user_info}
\title{Get basic information about the account of a specified user}
\usage{
get_user_info(user, accesstoken, verbose = FALSE, retry = TRUE)
}
\arguments{
\item{user}{Username of the Reddit user that is requested.}

\item{accesstoken}{The accesstoken required to access the endpoint. Scope
must be \code{"read"}.}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}

\item{retry}{A logical flag whether a failed api request should be retried.
Requests will be tried up to three times with varying time intervals between
requests.}
}
\value{
A dataframe with information about a specified user.
}
\description{
Get basic information about the account of a specified user
}
\examples{
\dontrun{
user_info <- get_user_info(user = "PresidentObama",
                           accesstoken = read_token)
}
}
