% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_wiki}
\alias{get_wiki}
\title{Get wiki pages for a specified subreddit.}
\usage{
get_wiki(
  subreddit,
  page = "all",
  accesstoken,
  output = c("df", "json", "all"),
  verbose = FALSE,
  retry = FALSE
)
}
\arguments{
\item{subreddit}{Name of the subreddit for which wiki pages are requested.}

\item{page}{Name of the requested page. If \code{"page"} is \code{"all"},
all wiki pages for the subreddit are returned.}

\item{accesstoken}{The access token required to access the endpoint. Scope of
the access token must be \code{"wikiread"}.}

\item{output}{What the function should return. Choose \code{json} for an 
unparsed json object, \code{df} for a parsed object in form of a dataframe, 
and \code{all} for a list containg the json object, a dataframe, and the 
before and after anchors (if those exist for the endpoint).}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}

\item{retry}{A logical flag whether a failed api request should be retried.
Requests will be tried up to three times with varying time intervals between
requests.}
}
\value{
A list containg the text of the page and all revisions.
If  \code{"page"} is \code{"all"}, a character vector containg the names of
all wiki pages.
}
\description{
Get wiki pages for a specified subreddit.
}
\examples{
\dontrun{
wiki_token <- get_token(scope = "wikiread",
                        useragent = my_useragent,
                        username = my_username,
                        password = my_password,
                        client_id = my_client_id,
                        client_secret = my_client_secret)

#All wiki pages of a subreddit
soccer_wiki <- get_wiki(subreddit = "soccer",
                       page = "all",
                       accesstoken = wiki_token)

#Access a specific wiki page
soccer_wiki_page <- get_wiki(subreddit = "soccer",
                            page = soccer_wiki[1,1],
                            accesstoken = wiki_token)
}
}
\seealso{
\code{\link{get_subreddit_info}}
}
