% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_functions.R
\name{get_wiki}
\alias{get_wiki}
\title{Get wiki pages for a specified subreddit.}
\usage{
get_wiki(subreddit = NULL, page = "all", accesstoken = NULL,
  verbose = FALSE)
}
\arguments{
\item{subreddit}{Name of the subreddit for which wiki pages are requested.}

\item{page}{Name of the requested page. If \code{"page"} is \code{"all"},
all wiki pages for the subreddit are returned.}

\item{accesstoken}{The access token required to access the endpoint. Scope of
the access token must be \code{"wikiread"}.}

\item{verbose}{A logical flag whether information about the data extraction
should be printed to the console.}
}
\value{
A list containg the text of the page and all revisions.
If  \code{"page"} is \code{"all"}, a character vector containg the names of
all wiki pages.
}
\description{
Get wiki pages for a specified subreddit.
}
\seealso{
\code{\link{get_subreddit_info}}
}
